Replaced the fixtures in test file
