+++ tests/conftest.py
@@ -5,7 +5,7 @@ from enum import Enum
 import pytest
 
 import nidaqmx.system
-from nidaqmx.constants import ExcitationSource, ProductCategory, RTDType, UsageTypeAI
+from nidaqmx.constants import ProductCategory, UsageTypeAI
 
 
 class Error(Exception):
@@ -202,61 +202,3 @@ def persisted_task(request):
 def test_assets_directory() -> pathlib.Path:
     """Gets path to test_assets directory."""
     return pathlib.Path(__file__).parent / "test_assets"
-
-
-@pytest.fixture(scope="module")
-def ai_voltage_chan_with_excit(any_x_series_device):
-    """Creates AI Channel object to measure Voltage.
-
-    Used in testing channel property of uint32 and boolean data types.
-    """
-    with nidaqmx.Task() as task:
-        ai_channel = task.ai_channels.add_ai_voltage_chan_with_excit(
-            any_x_series_device.ai_physical_chans[0].name,
-            voltage_excit_source=ExcitationSource.EXTERNAL,
-            voltage_excit_val=0.1,
-        )
-        yield ai_channel
-
-
-@pytest.fixture(scope="module")
-def ai_power_chan(sim_ts_power_device):
-    """Creates AI Channel object to measure Power.
-
-    Used in testing channel property of enum data type.
-    """
-    with nidaqmx.Task() as task:
-        ai_pwr_channel = task.ai_channels.add_ai_power_chan(
-            f"{sim_ts_power_device.name}/power",
-            voltage_setpoint=6.0,
-            current_setpoint=3.0,
-            output_enable=True,
-        )
-        yield ai_pwr_channel
-
-
-@pytest.fixture(scope="module")
-def ai_rtd_chan(any_x_series_device):
-    """Creates AI Channel object that use an RTD to measure temperature.
-
-    Used in testing channel property of float data type.
-    """
-    with nidaqmx.Task() as task:
-        ai_channel = task.ai_channels.add_ai_rtd_chan(
-            any_x_series_device.ai_physical_chans[0].name,
-            rtd_type=RTDType.PT_3750,
-        )
-        yield ai_channel
-
-
-@pytest.fixture(scope="module")
-def ci_pulse_width_chan(any_x_series_device):
-    """Creates CI Channel object to measure the width of a digital pulse.
-
-    Used in testing channel property of string data type.
-    """
-    with nidaqmx.Task() as task:
-        ci_channel = task.ci_channels.add_ci_pulse_width_chan(
-            any_x_series_device.ci_physical_chans[0].name,
-        )
-        yield ci_channel
